Moved AssetTokenCache from 'utils' to 'asset' directory
